+++ packages/connect-multichain/src/multichain/index.ts
@@ -61,12 +61,20 @@ import { DefaultTransport } from './transports/default';
 import { MultichainApiClientWrapperTransport } from './transports/multichainApiClientWrapper';
 import { MWPTransport } from './transports/mwp';
 import { keymanager } from './transports/mwp/KeyManager';
-import { getDappId, openDeeplink, setupDappMetadata } from './utils';
+import {
+  getDappId,
+  getGlobalObject,
+  mergeRequestedSessionWithExisting,
+  openDeeplink,
+  setupDappMetadata,
+} from './utils';
 
 export { getInfuraRpcUrls } from '../domain/multichain/api/infura';
 
 // ENFORCE NAMESPACE THAT CAN BE DISABLED
 const logger = createLogger('metamask-sdk:core');
+
+const SINGLETON_KEY = '__METAMASK_CONNECT_MULTICHAIN_SINGLETON__';
 
 export class MetaMaskConnectMultichain extends MultichainCore {
   readonly #provider: MultichainApiClient<RPCAPI>;
@@ -152,19 +160,49 @@ export class MetaMaskConnectMultichain extends MultichainCore {
     });
   }
 
+  // Creates a singleton instance of MetaMaskConnectMultichain.
+  // If the singleton already exists, it merges the incoming options with the
+  // existing singleton options for the following keys: `api.supportedNetworks`,
+  // `ui.*`, `mobile.*`, `transport.extensionId`, `debug`. Take note that the
+  // value for `dapp` is not merged as it does not make sense for subsequent calls to
+  // `createMultichainClient` to have a different `dapp` value.
   static async create(
     options: MultichainOptions,
   ): Promise<MetaMaskConnectMultichain> {
-    const instance = new MetaMaskConnectMultichain(options);
-    const isEnabled = await isLoggerEnabled(
-      'metamask-sdk:core',
-      instance.options.storage,
-    );
-    if (isEnabled) {
-      enableDebug('metamask-sdk:core');
+    const globalObject = getGlobalObject();
+    const existing = globalObject[SINGLETON_KEY] as
+      | Promise<MetaMaskConnectMultichain>
+      | undefined;
+    if (existing) {
+      const instance = await existing;
+      instance.mergeOptions(options);
+      if (options.debug) {
+        enableDebug('metamask-sdk:*');
+      }
+      return instance;
     }
-    await instance.#init();
-    return instance;
+
+    const instancePromise = (async (): Promise<MetaMaskConnectMultichain> => {
+      const instance = new MetaMaskConnectMultichain(options);
+      const isEnabled = await isLoggerEnabled(
+        'metamask-sdk:core',
+        instance.options.storage,
+      );
+      if (isEnabled) {
+        enableDebug('metamask-sdk:core');
+      }
+      await instance.#init();
+      return instance;
+    })();
+
+    globalObject[SINGLETON_KEY] = instancePromise;
+
+    instancePromise.catch((error) => {
+      globalObject[SINGLETON_KEY] = undefined;
+      console.error('Error initializing MetaMaskConnectMultichain', error);
+    });
+
+    return instancePromise;
   }
 
   async #setupAnalytics(): Promise<void> {
@@ -215,7 +253,7 @@ export class MetaMaskConnectMultichain extends MultichainCore {
         if (hasExtensionInstalled) {
           const apiTransport = new DefaultTransport();
           this.#transport = apiTransport;
-          this.#providerTransportWrapper.setupNotifcationListener();
+          this.#providerTransportWrapper.setupTransportNotificationListener();
           this.#listener = apiTransport.onNotification(
             this.#onTransportNotification.bind(this),
           );
@@ -227,7 +265,7 @@ export class MetaMaskConnectMultichain extends MultichainCore {
         const apiTransport = new MWPTransport(dappClient, kvstore);
         this.#dappClient = dappClient;
         this.#transport = apiTransport;
-        this.#providerTransportWrapper.setupNotifcationListener();
+        this.#providerTransportWrapper.setupTransportNotificationListener();
         this.#listener = apiTransport.onNotification(
           this.#onTransportNotification.bind(this),
         );
@@ -260,25 +298,16 @@ export class MetaMaskConnectMultichain extends MultichainCore {
 
   async #init(): Promise<void> {
     try {
-      // @ts-expect-error mmsdk should be accessible
-      if (typeof window !== 'undefined' && window.mmsdk?.isInitialized) {
-        logger('MetaMaskSDK: init already initialized');
-      } else {
-        await this.#setupAnalytics();
-        await this.#setupTransport();
-        try {
-          const baseProps = await getBaseAnalyticsProperties(
-            this.options,
-            this.storage,
-          );
-          analytics.track('mmconnect_initialized', baseProps);
-        } catch (error) {
-          logger('Error tracking initialized event', error);
-        }
-        if (typeof window !== 'undefined') {
-          // @ts-expect-error mmsdk should be accessible
-          window.mmsdk = this;
-        }
+      await this.#setupAnalytics();
+      await this.#setupTransport();
+      try {
+        const baseProps = await getBaseAnalyticsProperties(
+          this.options,
+          this.storage,
+        );
+        analytics.track('mmconnect_initialized', baseProps);
+      } catch (error) {
+        logger('Error tracking initialized event', error);
       }
     } catch (error) {
       await this.storage.removeTransport();
@@ -314,7 +343,7 @@ export class MetaMaskConnectMultichain extends MultichainCore {
     this.#dappClient = dappClient;
     const apiTransport = new MWPTransport(dappClient, kvstore);
     this.#transport = apiTransport;
-    this.#providerTransportWrapper.setupNotifcationListener();
+    this.#providerTransportWrapper.setupTransportNotificationListener();
     this.#listener = this.transport.onNotification(
       this.#onTransportNotification.bind(this),
     );
@@ -535,7 +564,7 @@ export class MetaMaskConnectMultichain extends MultichainCore {
       this.#onTransportNotification.bind(this),
     );
     this.#transport = transport;
-    this.#providerTransportWrapper.setupNotifcationListener();
+    this.#providerTransportWrapper.setupTransportNotificationListener();
     return transport;
   }
 
@@ -683,8 +712,13 @@ export class MetaMaskConnectMultichain extends MultichainCore {
     sessionProperties?: SessionProperties,
     forceRequest?: boolean,
   ): Promise<void> {
-    if (this.status !== 'connected') {
-      await this.disconnect();
+    if (
+      this.status === 'connecting' &&
+      this.transportType === TransportType.MWP
+    ) {
+      throw new Error(
+        'Existing connection is pending. Please check your MetaMask Mobile app to continue.',
+      );
     }
     const { ui } = this.options;
     const platformType = getPlatformType();
@@ -724,19 +758,29 @@ export class MetaMaskConnectMultichain extends MultichainCore {
       logger('Error tracking connection_initiated event', error);
     }
 
+    const sessionData = await this.#getCaipSession();
+
+    const { mergedScopes, mergedCaipAccountIds, mergedSessionProperties } =
+      mergeRequestedSessionWithExisting(
+        sessionData,
+        scopes,
+        caipAccountIds,
+        sessionProperties,
+      );
+
     // Needed because empty object will cause wallet_createSession to return an error
-    const nonEmptySessionProperites =
-      Object.keys(sessionProperties ?? {}).length > 0
-        ? sessionProperties
+    const nonEmptySessionProperties =
+      Object.keys(mergedSessionProperties ?? {}).length > 0
+        ? mergedSessionProperties
         : undefined;
 
     if (this.#transport?.isConnected() && !secure) {
       return this.#handleConnection(
         this.#transport
           .connect({
-            scopes,
-            caipAccountIds,
-            sessionProperties: nonEmptySessionProperites,
+            scopes: mergedScopes,
+            caipAccountIds: mergedCaipAccountIds,
+            sessionProperties: nonEmptySessionProperties,
             forceRequest,
           })
           .then(async () => {
@@ -755,9 +799,9 @@ export class MetaMaskConnectMultichain extends MultichainCore {
       const defaultTransport = await this.#setupDefaultTransport();
       return this.#handleConnection(
         defaultTransport.connect({
-          scopes,
-          caipAccountIds,
-          sessionProperties: nonEmptySessionProperites,
+          scopes: mergedScopes,
+          caipAccountIds: mergedCaipAccountIds,
+          sessionProperties: nonEmptySessionProperties,
           forceRequest,
         }),
         scopes,
@@ -771,9 +815,9 @@ export class MetaMaskConnectMultichain extends MultichainCore {
       // Web transport has no initial payload
       return this.#handleConnection(
         defaultTransport.connect({
-          scopes,
-          caipAccountIds,
-          sessionProperties: nonEmptySessionProperites,
+          scopes: mergedScopes,
+          caipAccountIds: mergedCaipAccountIds,
+          sessionProperties: nonEmptySessionProperties,
           forceRequest,
         }),
         scopes,
@@ -793,9 +837,9 @@ export class MetaMaskConnectMultichain extends MultichainCore {
       // Desktop is not preferred option, so we use deeplinks (mobile web)
       return this.#handleConnection(
         this.#deeplinkConnect(
-          scopes,
-          caipAccountIds,
-          nonEmptySessionProperites,
+          mergedScopes,
+          mergedCaipAccountIds,
+          nonEmptySessionProperties,
         ),
         scopes,
         transportType,
@@ -806,9 +850,9 @@ export class MetaMaskConnectMultichain extends MultichainCore {
     return this.#handleConnection(
       this.#showInstallModal(
         shouldShowInstallModal,
-        scopes,
-        caipAccountIds,
-        nonEmptySessionProperites,
+        mergedScopes,
+        mergedCaipAccountIds,
+        nonEmptySessionProperties,
       ),
       scopes,
       transportType,
@@ -820,20 +864,47 @@ export class MetaMaskConnectMultichain extends MultichainCore {
     super.emit(event, args);
   }
 
-  async disconnect(): Promise<void> {
-    await this.#listener?.();
-    this.#beforeUnloadListener?.();
+  async #getCaipSession(): Promise<SessionData> {
+    let sessionData: SessionData = {
+      sessionScopes: {},
+      sessionProperties: {},
+    };
+    if (this.status === 'connected') {
+      const response = await this.transport.request({
+        method: 'wallet_getSession',
+      });
+      if (response.result) {
+        sessionData = response.result as SessionData;
+      }
+    }
+    return sessionData;
+  }
 
-    await this.#transport?.disconnect();
-    await this.storage.removeTransport();
+  async disconnect(scopes: Scope[] = []): Promise<void> {
+    const sessionData = await this.#getCaipSession();
 
-    this.emit('stateChanged', 'disconnected');
+    const remainingScopes =
+      scopes.length === 0
+        ? []
+        : Object.keys(sessionData.sessionScopes).filter(
+            (scope) => !scopes.includes(scope as Scope),
+          );
 
-    this.#listener = undefined;
-    this.#beforeUnloadListener = undefined;
-    this.#transport = undefined;
-    this.#providerTransportWrapper.clearNotificationCallbacks();
-    this.#dappClient = undefined;
+    await this.#transport?.disconnect(scopes);
+
+    if (remainingScopes.length === 0) {
+      await this.#listener?.();
+      this.#beforeUnloadListener?.();
+
+      await this.storage.removeTransport();
+
+      this.#listener = undefined;
+      this.#beforeUnloadListener = undefined;
+      this.#transport = undefined;
+      this.#providerTransportWrapper.clearTransportNotificationListener();
+      this.#dappClient = undefined;
+      this.status = 'disconnected';
+    }
   }
 
   async invokeMethod(request: InvokeMethodOptions): Promise<Json> {
@@ -867,5 +938,27 @@ export class MetaMaskConnectMultichain extends MultichainCore {
         }
       }, 10); // small delay to ensure the message encryption and dispatch completes
     }
+  }
+
+  // Provides a way for ecosystem clients (EVM, Solana, etc.) to get the current CAIP session data
+  // when instantiating themselves (as they would have already missed any initial sessionChanged events emitted by ConnectMultichain)
+  // without having to concern themselves with the current transport connection status.
+  async emitSessionChanged(): Promise<void> {
+    const emptySession = { sessionScopes: {} };
+
+    if (this.status !== 'connected' && this.status !== 'connecting') {
+      // If we aren't connected or connecting, there definitely is no active CAIP session
+      // so we optimistically emit an empty session to signify that to the ecosystem client consumers (EVM, Solana, etc.)
+      this.emit('wallet_sessionChanged', emptySession);
+      return;
+    }
+
+    // Otherwise, we need to fetch the current CAIP session from the wallet
+    const response = await this.transport.request({
+      method: 'wallet_getSession',
+    });
+
+    // And then simulate a sessionChanged event with the current CAIP session data
+    this.emit('wallet_sessionChanged', response.result ?? emptySession);
   }
 }